+++ CHANGES.rst
@@ -15,6 +15,8 @@ Changes:
 - Add support of `HTML` responses for `OGC API - Processes` endpoints
   (fixes `#210 <https://github.com/crim-ca/weaver/issues/210>`_).
 - Add ``weaver.wps_restapi_html`` configuration setting to control support of `HTML` responses.
+- Add ``weaver.wps_restapi_html_override_user_agent`` configuration setting for control of default `HTML` or `JSON`
+  rendering by requests from web browsers.
 - Refactor ``pyramid`` configuration to employ ``Configurator.add_cornice_service``
   utility instead of ``Configurator.add_route`` and ``Configurator.add_view`` handlers that were causing a lot of
   duplication between the ``cornice.Service`` parametrization and their corresponding view decorators. All metadata